feat: change the cmd layout for go install
